+++ +sbfsem/+analysis/AnnotationSizes.m
@@ -1,4 +1,20 @@
 classdef AnnotationSizes < sbfsem.analysis.NeuronAnalysis
+% ANNOTATIONSIZES
+% 
+% Description:
+%	A histogram of dendrite radius as a function of distance from soma
+% 
+% Methods:
+% 	See NeuronAnalysis for information on using the core 3 methods:
+%		constructor, doAnalysis, plot
+%	and for accessing the specific key/value arguments with `help`
+%
+% See also:
+%	SomaDistanceView, sbfsem.analysis.NeuronAnalysis
+%
+% History:
+%	18Sept2018 - SSP
+% --------------------------------------------------------------------
 
 	properties (Constant = true, Hidden = true)
 		DisplayName = 'AnnotationSizes';
@@ -14,6 +30,19 @@ classdef AnnotationSizes < sbfsem.analysis.NeuronAnalysis
 		end
 
 		function doAnalysis(obj, varargin)
+			% DOANALYSIS
+			%	
+			% Optional key/value inputs:
+			%	nBins 		Number of histogram bins 
+			% 				default lets matlab histcounts decide
+			%	Plot 		Plot the output? (Default = true)
+			%	Max 		Max annotation size to include (default = all)
+			%				Do this to zoom in on a certain range
+			%
+			% Example:
+			%	% Analyze only the annotations with > 3 micron radius
+			%	obj.doAnalysis('Max', 3);
+			
 			ip = inputParser();
 			ip.CaseSensitive = false;
 			addParameter(ip, 'nBins', [], @isnumeric);
@@ -45,6 +74,15 @@ classdef AnnotationSizes < sbfsem.analysis.NeuronAnalysis
 		end
 
 		function plot(obj, varargin)
+			% PLOT
+			%
+			% Optional key/value inputs:
+			%	useDiameter 	Plot diameter instead of radius
+			%					Default = false
+			%	ax 				Axis handle to plot to 
+			%					Default = new figure
+			%	color 			Color to plot as (default = black)
+
 			ip = inputParser();
 			ip.CaseSensitive = false;
 			addParameter(ip, 'ax', [], @ishandle);